GEN-26 API issue creator for Jira (#5)
It should probably not be a surprise that working with
the Jira API is about as difficult as working with the
Jira interface. The API calls have to match the screen
fields for issues/bugs/etc on particular boards, or the
call will fail hard.

There is probably a consultant gig in here somewhere.