GEN-15 batch download all issues from a project (#25)
This the next step for getting all issues into S3,
namely, being able to get all the issues in a project
in batches of 50, which is the maximum size allowed by
Jira. In this change, the issues are written to /tmp/jira.
In the future, the collection of downloaded issues should
be passed to an async method which uploads to S3. For now,
everything is as simple as possible, as the script only
needs to be run intermittently, not continuously.