GEN-64 Factor S3 uploading into dedicated class (#26)
S3 operations are now moved to their own class.

Specs are started but deficient. It's not clear
how to unit test. "Integration" testing is by
running the script and checking the S3 bucket for
timestamps.

The end goal is to have a script which only configures,
then invokes an execution. The script should have no logic
in it at all.