GEN-74 Upload all Jira issues to S3
"Make the change easy then make the easy change"

This is a great example of the above. Getting all the Jira
issues for a project has to be done in batches of maximum 50.
Previous commits worked out the structure for how this ought
to work. This commit implements batch downloading for all the
projects, passing in where the issues are written as a `writer`
callback.