GEN-88 runbook for gem updating (#36)
This is a little detour, it's a procrastination while
avoiding harder tasks.

However, it will be useful and save some time for creating
gem updating tickets. That is, it saves time for projects
which are serious about keeping gems up to date. Projects
which never update gems won't notice the difference.